rec: Record based on time and no auto-normalization
Amplitude detection didn't work too well for me. :/

rec now keeps 1/32s of history. As soon as we detect a key, that history
will be written to the output file. After that, another 1/2s will be
recorded to catch all the "zoingggggg" noises.

Auto-normalization was a bad idea to begin with. We must find the global
peak value of all samples and then normalize all of them using the same
gain.