Add desktop file
OK, since it clear that systemd is not a way to start buckler, there is
a buckler.desktop file. It could be putted into /etc/xdg/autostart/ or
~/.config/autostart/ directories. So buckler will start with user
session start. No addition actions needed.

The relative disadvantage compared to systemd service is that if buckle
will crash it will not started again.